[JetBrains] bump up latest backend plugin platform version and fix integration tests (#20109)
* Update Platform Version of JetBrains Backend Plugin (EAP) to

* Revert unnecessary changes

* Fix project name and build target name

* Fix integration tests

---------

+++ dev/jetbrains-test/test.sh
@@ -4,6 +4,11 @@ GATEWAY_PLUGIN_PATH=$(pwd)/gitpod-gateway.zip
 export GATEWAY_PLUGIN_PATH
 HOME=/home/gitpod
 
+if [ ! -f "$GATEWAY_PLUGIN_PATH" ]; then
+  echo "Gateway plugin zip not found at $GATEWAY_PLUGIN_PATH"
+  exit 1
+fi
+
 mkdir -p $HOME/.local/share/JetBrains/consentOptions/
 echo -n "rsch.send.usage.stat:1.1:0:1644945193441" > $HOME/.local/share/JetBrains/consentOptions/accepted
 mkdir -p $HOME/.config/JetBrains/JetBrainsClient/options
